% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpe.R
\name{display}
\alias{display}
\title{perform}
\usage{
display(data, linpe)
}
\arguments{
\item{data}{a data frame object}

\item{linpe}{the name of the linpe attribute previously saved with `link()`}
}
\value{
NULL and, as side effect, opens the `.rmd` file corresponding to the analysis
}
\description{
perform
}
\examples{
data(mtcars_linpe)
display(mtcars_linpe, linpe = "test-linpe")
}

