% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpe.R
\name{enumerate}
\alias{enumerate}
\title{enumerate}
\usage{
enumerate(data)
}
\arguments{
\item{data}{a data frame object}
}
\value{
list of linpes associated to data
}
\description{
enumerate
}
\examples{
data(mtcars_linpe)
enumerate(mtcars_linpe)
enumerate(mtcars)
}

