% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpe.R
\name{link}
\alias{link}
\title{link}
\usage{
link(data, file, linpe = NULL, sep = "\\n")
}
\arguments{
\item{data}{a data frame object}

\item{file}{Character vector of length one pointing to a .rmd file location}

\item{linpe}{Character vector of length one giving the name of the linpe (analysis). When `linpe=NULL` (default value),
the name of the linpe is given  by the file name}

\item{sep}{end of line character in teh `.rmd` file file. Default to "\n"}
}
\value{
A dataframe with an analysis attribute linked to it
}
\description{
link
}
\examples{
file <- paste(find.package("linpe"), "rmd/test-linpe.rmd", sep = "/")
# analysis is saved with default name
mtcars_linpe <- link (mtcars, file = file)
attr(mtcars_linpe, "test-linpe")

# analysis is saved with specific name
mtcars_linpe <- link (mtcars, file, analysis = "this-analysis")
attr(mtcars_linpe, "this-analysis")
}

