% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpe.R
\name{linpe}
\alias{linpe}
\title{linpe}
\usage{
linpe(data)
}
\arguments{
\item{data}{a data frame object}
}
\value{
list of linpes associated to data
}
\description{
linpe
}
\examples{
data(mtcars_linpe)
linpe(mtcars_linpe)
linpe(mtcars)
}

