% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpe.R
\name{perform}
\alias{perform}
\title{perform}
\usage{
perform(data, linpe, ...)
}
\arguments{
\item{data}{a data framae object}

\item{linpe}{the name of an `linpe` attribute previously saved with `link()`}

\item{...}{any parameter to be passed to function rmarkdown::render}
}
\value{
invisible(NULL) and, as side effect, renders the `.rmd` file corrensponding to the analysis and display it in the default html viewer
}
\description{
perform
}
\examples{
data(mtcars_linpe)
perform(mtcars_linpe, "test-linpe")
}

