% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpe.R
\name{unlink}
\alias{unlink}
\title{unlink}
\usage{
unlink(data, linpe)
}
\arguments{
\item{data}{a data frame object}

\item{linpe}{Character vector of length one giving the name of the linpe (analysis)}
}
\value{
A dataframe with a `linpe` attribute removed from it
}
\description{
unlink
}
\examples{
file <- paste(find.package("linpe"), "rmd/test-linpe.rmd", sep = "/")
mtcars_linpe <- link (mtcars, file = file)
linpe(mtcars_linpe)
mtcars_linpe <- unlink (mtcars_linpe, linpe = "test-linpe")
linpe(mtcars_linpe)

}

