% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpe.R
\name{display}
\alias{display}
\title{display}
\usage{
display(data, linpe)
}
\arguments{
\item{data}{a data frame object}

\item{linpe}{the name of the linpe attribute previously saved with `link()`}
}
\value{
invisible(NULL) and, as side effect, opens the `.rmd` linked to the linpe
}
\description{
display
}
\examples{
data(mtcars_linpe)
display(mtcars_linpe, linpe = "test-linpe")
}

